server.py
import os
import json
import time
import shutil
from flask import Flask, request, jsonify, abort, render_template, send_from_directory
from flask_cors import CORS
from werkzeug.utils import secure_filename
from dotenv import load_dotenv
# (O resto da sua lógica de backend que lida com o genai, chats, arquivos, etc.)
# ...

app = Flask(__name__, template_folder='templates', static_folder='static')
CORS(app)

# --- ROTA PARA SERVIR A INTERFACE (HTML) ---
@app.route('/')
def index():
    return render_template('index.html')

# --- SUAS ROTAS DE API EXISTENTES ---
# Ex: @app.route('/api/assistants', methods=['GET'])
# ... (cole aqui todas as suas rotas de /api/... que estavam no app2.py)

if __name__ == '__main__':
    # initialize_data_structure() # Garanta que esta função exista se for necessária
    app.run(host='0.0.0.0', port=int(os.environ.get('PORT', 8080)))