requirements.txt
Flask
Flask-Cors
python-dotenv
google-generativeai
werkzeug
gunicorn # Adicionado para produção na nuvem